+++ packages/aws-amplify-angular/src/__tests__/components/authenticator/sign-up.component.ionic.spec.ts
@@ -1,0 +1,40 @@
+import { Component } from '@angular/core';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
+import { AmplifyService } from '../../../providers/amplify.service'
+import { SignUpComponentIonic } from '../../../components/authenticator/sign-up-component/sign-up.component.ionic'
+
+
+describe('SignUpComponentCore: ', () => {
+
+  let component: SignUpComponentIonic;
+  let service: AmplifyService;
+
+  beforeEach(() => { 
+    service = new AmplifyService();
+    component = new SignUpComponentIonic(service);
+  });
+
+  afterEach(() => {
+    service = null;
+    component = null;
+  });
+
+
+  it('...should be created', () => {
+    expect(component).toBeTruthy();
+  });
+
+  it('...should have an onConfirmSignUp method', () => {
+    expect(component.onConfirmSignUp).toBeTruthy();
+  });
+
+  it('...should have an onSignIn method', () => {
+    expect(component.onSignIn).toBeTruthy();
+  });
+
+  it('...should have an onSignUp method', () => {
+    expect(component.onSignUp).toBeTruthy();
+  });
+
+});