+++ packages/aws-amplify-react/__tests__/Auth/Provider/withGoogle-test.js
@@ -1,18 +1,4 @@
-jest.mock('../../../src/Categories', () => {
-    const Auth = {
-        federatedSignIn() {
-            return;
-        },
-        currentAuthenticatedUser() {
-            return;
-        }
-    };
-
-    return {
-        Auth
-    };
-});
-import { Auth } from '../../../src/Categories';
+import Auth from '@aws-amplify/auth';
 import React, { Component } from 'react';
 import withGoogle, {
     GoogleButton