+++ packages/aws-amplify-react/src/Auth/Greetings.jsx
@@ -13,7 +13,7 @@
 
 import React, { Component } from 'react';
 import { I18n, Hub, ConsoleLogger as Logger } from '@aws-amplify/core';
-import { Auth } from '../Categories';
+import Auth from '@aws-amplify/auth';
 
 import AuthPiece from './AuthPiece';
 import { NavBar, Nav, NavRight, NavItem, NavButton } from '../AmplifyUI';
@@ -116,10 +116,7 @@ export default class Greetings extends AuthPiece {
             })
             .catch(err => {
                 if (!that._isMounted) { return; }
-                if (!authState || authState === 'signedIn') {
-                    this.setState({ authState: 'signIn' });
-                    this.changeState('signIn');
-                }
+                this.signOut();
             });
     }