+++ packages/core/CHANGELOG.md
@@ -3,6 +3,89 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.4"></a>
+## [1.0.4](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.3-unstable.1...@aws-amplify/core@1.0.4) (2018-08-06)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.3-unstable.1"></a>
+## [1.0.3-unstable.1](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.3...@aws-amplify/core@1.0.3-unstable.1) (2018-08-06)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.3-unstable.0...@aws-amplify/core@1.0.3) (2018-07-28)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.3-unstable.0"></a>
+## [1.0.3-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.2...@aws-amplify/core@1.0.3-unstable.0) (2018-07-26)
+
+
+### Bug Fixes
+
+* **aws-amplify:** use method get instead of getPromise for credentials ([f15d74f](https://github.com/aws/aws-amplify/commit/f15d74f))
+
+
+
+
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.2-unstable.0...@aws-amplify/core@1.0.2) (2018-07-19)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.2-unstable.0"></a>
+## [1.0.2-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1...@aws-amplify/core@1.0.2-unstable.0) (2018-07-19)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.1"></a>
+## [1.0.1](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1-unstable.2...@aws-amplify/core@1.0.1) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.1-unstable.2"></a>
+## [1.0.1-unstable.2](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1-unstable.1...@aws-amplify/core@1.0.1-unstable.2) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.1-unstable.1"></a>
+## [1.0.1-unstable.1](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1...@aws-amplify/core@1.0.1-unstable.1) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
+<a name="1.0.1-unstable.0"></a>
+## [1.0.1-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1...@aws-amplify/core@1.0.1-unstable.0) (2018-07-18)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/core
+
 <a name="1.0.1-next.2"></a>
 ## [1.0.1-next.2](https://github.com/aws/aws-amplify/compare/@aws-amplify/core@1.0.1-next.1...@aws-amplify/core@1.0.1-next.2) (2018-06-28)